fix: signed out state for AccountGate
